Split is_usda_disabled and is_usda_elderly into their own Python files
